+++ Controls-demo/gridNew/ColumnSeparator/PartialColumnSeparator/VariantThree/VariantThree.wml
@@ -1,0 +1,17 @@
+<div class="controlsDemo__wrapper">
+    <div class="controlsDemo__inline-block" style="max-width: 620px; margin-right: 20px; margin-bottom: 20px;">
+        <div style="margin-bottom: 20px;">
+            <Controls.toggle:Checkbox bind:value="_rowSeparator" caption="Show row separator"/>
+            <Controls.toggle:Checkbox bind:value="_columnSeparator" caption="Show column separator"/>
+        </div>
+        <Controls.grid:View
+                keyProperty="key"
+                source="{{_viewSource}}"
+                header="{{_header}}"
+                columns="{{_columns}}"
+                multiSelectVisibility="visible"
+                rowSeparatorSize="{{ _rowSeparator ? 's' }}"
+                columnSeparatorSize="{{ _columnSeparator ? 's' }}">
+        </Controls.grid:View>
+    </div>
+</div>
